MOving Error to error.rs
